requirements.txt
Django==4.2.1
kafka-python==2.0.2
six==1.16.0